research-assistant/requirements.txt
# Core Dependencies
endee
sentence-transformers>=2.2.2
google-generativeai>=0.3.0
requests>=2.28.0

# Web Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
arxiv>=2.0.0
tiktoken>=0.5.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tenacity>=8.2.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
